Fix clevis/tang: support multiple phys. volumes
When multiple disks are available and `autopart` is used in the kickstart config,
all disks get encrypted and put into a volume group on which a logical volume
for the root partition is created. This means we need to recursively resolve all
underlying volumes for the root partition and setup tang for them.

As this fix should also work for Ubuntu, we also address some bugs
with the Ubuntu support:
- disk_enc_clevis_tang.erb needs to work with dash (do not use bash features)
- minor versions are empty for Ubuntu/Debian, so do not check for os_minor
- PKG_MANAGER_INSTALL is not defined in preseed_autoinstall_cloud_init.erb