Fix clevis/tang: support multiple phys. volumes
Often there is more than a single disk that gets encrypted during
installation. When multiple disks are available and `autopart` is used
in the kickstart config, all disks get encrypted and put into a volume
group on which a logical volume for the root partition is created.
Another example is when users add a separate disk for directories like
/var to separate application data. This means we need to recursively
resolve all underlying volumes of encrypted disks, not just for the
root partition. We will therefore do this for all entries in
/etc/crypttab, which should catch all disks that were encrypted in the
installation process.

As this feature should also work for Ubuntu, we also address some bugs
with the Ubuntu support:
- disk_enc_clevis_tang.erb needs to work with dash (not just bash)
- minor versions are empty for Ubuntu/Debian, do not check for os_minor
- PKG_MANAGER_INSTALL is missing in preseed_autoinstall_cloud_init.erb